feat: Creating private chat
We now have a distinction between a group chat and a private chat, although both of them will have the members_ entries.

In order for us to check if a user already have a private chat with another user, we are using the chatId being a combination between both usernames. Unfortunately we cannot sort the usernames, which would allow an easy way to see if there is already a chat.